requirements.txt
fastapi==0.85.0
uvicorn==0.18.3
sqlalchemy==1.4.41
python-multipart==0.0.5
passlib==1.7.4
bcrypt==4.0.0
pydantic==1.10.1
python-jose==3.3.0